[Validations] Restructure and optimizations
